notebooks/requirements.txt
aiohttp==3.13.3
azure-ai-projects==2.0.1
azure-identity==1.25.2
azure-search-documents==11.7.0b2
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.23.0
ipykernel==7.2.0
ipywidgets==8.1.8
openai==2.26.0
pandas==3.0.1
python-dotenv==1.2.2
requests==2.32.5
